make skill category rank editable

+++ models.py
@@ -22,7 +22,7 @@ class Project(models.Model):
 
 class SkillCategory(models.Model):
     name = models.CharField(max_length=20)
-    rank = models.AutoField(primary_key=True)
+    rank = models.PositiveSmallIntegerField(primary_key=True)
 
     def __str__(self):
         return self.name
